# swagger/swagger_autogen.json
{
  "swagger": "2.0",
  "info": {
    "title": "RISC V!",
    "description": "Risc V Test",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/health-check": {
      "get": {
        "tags": [
          "Risc V"
        ],
        "description": "Health check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/assemble-code": {
      "post": {
        "tags": [
          "Risc V"
        ],
        "description": "Assemble code",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "memories": {
                  "example": "any"
                },
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/run-one-step": {
      "post": {
        "tags": [
          "Risc V"
        ],
        "description": "Run one line",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "memories": {
                  "example": "any"
                },
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/run-entire-program": {
      "post": {
        "tags": [
          "Risc V"
        ],
        "description": "Run entire program",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "memories": {
                  "example": "any"
                },
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/dump-machine-code": {
      "post": {
        "tags": [
          "Risc V"
        ],
        "description": "Dump code",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "asm": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}